Generate similarity distribution plot

Generate similarity distribution plot

Usage

plot_simi_dist(df_simi, simi_param)

Arguments

df_simi

A data frame with string similarity results.

simi_param

A character of the similarity parameter to plot the distribution (i.e., quoted name of the column containing the similarity scores). For example, "title_simi", "abstract_simi", "author_simi", "first_author_last_name_simi".

Value

A scatterplot of distributions of similarity scores

Examples

if (FALSE) {
# `df_simi` is the data frame resulted from `simi_order_df()`
# see the example in the `simi_order_df()` help page for how `df_simi` is generated

# Distribution of similarity scores based on normalized title
p_ti <- plot_simi_dist(df_simi, "title_simi")
p_ti  # show p_ti in the Plots tab

# Distribution of similarity scores based on normalized abstract
p_ab <- plot_simi_dist(df_simi, "abstract_simi")
p_ab  # show p_ab in the Plots tab
}